Codecleaning, renaming, added TODOs.
