+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/DatabaseAccess.java
@@ -6,9 +6,11 @@ public class DatabaseAccess {
     // Workaround class to enable easy switching between different database implementations.
     // Use this class whenever access to the database is required.
 
+    // TODO: Use a proper factory for database creation.
+
 
     private static Database defaultDatabaseInstance;
-    private static boolean initialized = false;
+    private static boolean isInitialized = false;
 
 
     private static void createDefaultDatabaseInstance(Context appContext) {
@@ -22,16 +24,16 @@ public class DatabaseAccess {
     // Public interface:
 
     public static void init(Context appContext) {
-        if (initialized) {
+        if (isInitialized) {
             throw new StorageException("Attempt to reinitialize database.");
         }
         createDefaultDatabaseInstance(appContext);
-        initialized = true;
+        isInitialized = true;
     }
 
 
     public static Database getDefaultDatabaseInstance() {
-        if (!initialized) {
+        if (!isInitialized) {
             throw new StorageException("Cannot access database. Database not initialized yet.");
         }
         return defaultDatabaseInstance;
