+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/persistent/DaoTEK.java
@@ -1,4 +1,4 @@
-package org.coralibre.android.sdk.internal.database.persistenDatabase;
+package org.coralibre.android.sdk.internal.database.persistent;
 
 
 import androidx.room.Dao;
