+++ sdk/src/main/java/org/coralibre/android/sdk/internal/database/persistent/RoomDatabaseDelegate.java
@@ -1,4 +1,4 @@
-package org.coralibre.android.sdk.internal.database.persistenDatabase;
+package org.coralibre.android.sdk.internal.database.persistent;
 
 import androidx.room.RoomDatabase;
 import androidx.room.TypeConverters;
@@ -14,7 +14,7 @@ import org.coralibre.android.sdk.internal.database.model.entity.EntityGeneratedT
 @TypeConverters(
         {ENNumberConverter.class}
         )
-public abstract class RoomDatabaseAbstrImpl extends RoomDatabase {
+public abstract class RoomDatabaseDelegate extends RoomDatabase {
 
     public abstract DaoTEK daoTEK();
     public abstract DaoCapturedData daoCapturedData();
